# Enhance large files with syntax highlighting performance