@donho 2024-12-09
@softmgr 
Thank you for the code you've provided. 
Could you review this PR and test the binary please?

I tested with 400 MB large XML file, put the position in the middle, then quit Notepad++ and relaunch it. 
The content is loaded rapidly with the position set previously and without styling (yet). Few seconds later, the document is styled completely - which is good.

If I switch off this large file, by keeping its position, for switching in to a small file, close Notepad++, and relaunch it, the small file is displayed as expected. Now I switch back to this large XML file, same thing happens as described above (the content is loaded immediately and the right position scrolled, the styling comes after) - so far so good.

However, without quitting this Notepad++ instance, when the large XML file is switched off then switched back, it's frozen few seconds then the styled content is shown. 

I didn't find what I've done wrong, so you might have the idea about it.

Note that the PR in its first stage, It could be modified during the code review.