@softmgr 2024-12-10
According to my personal analysis, the factors affecting the performance of large files are mainly the following three:
1.	Syntax highlighting;
2.	Excessive number of folded lines;
3.	Auto-completion.

Additionally, I suggest removing the `Clickable Link` feature, as it slows down the overall performance of the editor. (We can use the right-click menu to copy and open links. If you’re interested, I can provide my improved code for reference.)

The following code is based on the `2024-12-10` development version and is for reference only:

1. Improvements to Syntax Highlighting:	Solves issues such as lag when loading large files, switching between large files, and moving the cursor within large files.


// **ScintillaEditView.h**

```
#define MODEVENTMASK_ON 3  //SC_MOD_INSERTTEXT | SC_MOD_DELETETEXT
#define MAX_FOLD_LINES_MORE_THAN 99  // When the number of rows collapsed in bulk is greater than this value, the WM_SETREDRAW message is blocked

// ...
class ScintillaEditView : public Window
{
friend class Finder;
public:
	// ...
	virtual void destroy()
	{
		if (_blankDocument != 0) // added
		{
			execute(SCI_RELEASEDOCUMENT, 0, _blankDocument); // added
			_blankDocument = 0; // added
		}
		::DestroyWindow(_hSelf);
		_hSelf = NULL;
		_pScintillaFunc = NULL;
	};
	// ...
	std::pair<size_t, size_t> getSelectionLinesRange(intptr_t selectionNumber = -1) const;
	std::pair<Sci_Position, Sci_Position> getSelectionPosition(intptr_t selectionNumber = -1, int sort = 1) const; // added

	// ...
	Document getBlankDocument(/*bool reset = false*/); // added

protected:
	// ...
	Buffer* _prevBuffer = nullptr; // added
	Document _blankDocument = 0; // added
	// ...
};
```


// **ScintillaEditView.cpp**

```
void ScintillaEditView::init(HINSTANCE hInst, HWND hPere)
{
	// ...
	if (!_pScintillaPtr)
	{
		throw std::runtime_error("ScintillaEditView::init : SCI_GETDIRECTPOINTER message failed");
	}

	execute(SCI_SETMODEVENTMASK, 0); // added
	execute(SCI_SETIDLESTYLING, SC_IDLESTYLING_ALL, 0); // added
	execute(SCI_SETMARGINMASKN, _SC_MARGE_FOLDER, SC_MASK_FOLDERS);
	
	// ...
	if (_defaultCharList.empty())
	{
		// ...
	}
	execute(SCI_SETMODEVENTMASK, MODEVENTMASK_ON); // added
	//Get the startup document and make a buffer for it so it can be accessed like a file
	attachDefaultDoc();
}

void ScintillaEditView::activateBuffer(BufferID buffer, bool force)
{
	// ...

	// put the state into the future ex buffer
	_currentBuffer->setHeaderLineState(lineStateVector, this);

	_prevBuffer = _currentBuffer; // added
	
	_currentBufferID = buffer;	//the magical switch happens here
	_currentBuffer = newBuf;
	
	// ...

	////////////////////////////////////////////////////////////////////
	const bool isSameLangType = _prevBuffer != nullptr && ((_prevBuffer == _currentBuffer) || (_prevBuffer->getLangType() == _currentBuffer->getLangType()));
	const int currentLangInt = static_cast<int>(_currentBuffer->getLangType());
	const bool isFirstActiveBuffer = (_currentBuffer->getLastLangType() != currentLangInt) || (_currentBuffer->isUntitled());
	if (!isSameLangType && !isFirstActiveBuffer)  // When entering the tab for the second or more times
	{
		execute(SCI_SETMODEVENTMASK, 0);  // Turn OFF the notifications
		execute(SCI_SETDOCPOINTER, 0, getBlankDocument());
		execute(SCI_SETMODEVENTMASK, MODEVENTMASK_ON);  // Turn ON the notifications
		defineDocType(_currentBuffer->getLangType());
	}
	////////////////////////////////////////////////////////////////////

	// change the doc, this operation will decrease
	// the ref count of old current doc and increase the one of the new doc. FileManager should manage the rest
	// Note that the actual reference in the Buffer itself is NOT decreased, Notepad_plus does that if neccessary
	execute(SCI_SETMODEVENTMASK, 0);  // Turn OFF the notifications
	execute(SCI_SETDOCPOINTER, 0, _currentBuffer->getDocument());
	execute(SCI_SETMODEVENTMASK, MODEVENTMASK_ON);  // Turn ON the notifications

	// Due to execute(SCI_CLEARDOCUMENTSTYLE); in defineDocType() function
	// defineDocType() function should be called here, but not be after the fold info loop
	if (isFirstActiveBuffer)  // This is necessary when entering the tab for the first time
		defineDocType(_currentBuffer->getLangType());
	
	_currentBuffer->setLastLangType(currentLangInt); // added

	setWordChars();

	//deleted code (It doesn't seem necessary to call it here?)
	//if (_currentBuffer->getNeedsLexing())
	//{
	//	restyleBuffer();
	//}

	// ...
	
	//deleted code (It doesn't seem necessary to call it here?)
	//runMarkers(true, 0, true, false);

	// ...
}

void ScintillaEditView::defineDocType(LangType typeDoc)
{
	// ...
	if (svp._indentGuideLineShow)
	{
		// ...
	}
	
    execute(SCI_SETLAYOUTCACHE, SC_CACHE_DOCUMENT, 0); // added
	execute(SCI_STARTSTYLING, 0, 0); // added
}

Document ScintillaEditView::getBlankDocument(/*bool reset*/)
{
	//if (reset)
	//{
	//	if (_blankDocument != 0)
	//	{
	//		execute(SCI_RELEASEDOCUMENT, 0, _blankDocument);
	//		_blankDocument = 0;
	//	}
	//}
	if (_blankDocument == 0)
	{
		_blankDocument = static_cast<Document>(execute(SCI_CREATEDOCUMENT, 0, SC_DOCUMENTOPTION_TEXT_LARGE));
		execute(SCI_ADDREFDOCUMENT, 0, _blankDocument);
	}
	return _blankDocument;
}

void ScintillaEditView::saveCurrentPos()
{
	//Save data so, that the current topline becomes visible again after restoring.
	size_t displayedLine = execute(SCI_GETFIRSTVISIBLELINE);
	size_t docLine = execute(SCI_DOCLINEFROMVISIBLE, displayedLine);		//linenumber of the line displayed in the top
	size_t offset = displayedLine - execute(SCI_VISIBLEFROMDOCLINE, docLine);		//use this to calc offset of wrap. If no wrap this should be zero
	size_t wrapCount = execute(SCI_WRAPCOUNT, docLine);

	Buffer * buf = MainFileManager.getBufferByID(_currentBufferID);

	Position pos;
	// the correct visible line number
	pos._firstVisibleLine = docLine;
	//deletetd code:  //pos._startPos = execute(SCI_GETANCHOR);
	//deletetd code:  //pos._endPos = execute(SCI_GETCURRENTPOS);
	pos._xOffset = execute(SCI_GETXOFFSET);
	pos._selMode = execute(SCI_GETSELECTIONMODE);
	pos._scrollWidth = execute(SCI_GETSCROLLWIDTH);
	pos._offset = offset;
	pos._wrapCount = wrapCount;

	const intptr_t nbSelections = execute(SCI_GETSELECTIONS);
	if (pos._selMode == SC_SEL_STREAM && nbSelections > 1)
	{
		const intptr_t startSelection = (nbSelections <= 99) ? 0 : nbSelections - 99;
		for (intptr_t i = startSelection; i < nbSelections; ++i)
		{
			pos._selections.emplace_back(getSelectionPosition(i, 0));
		}
	}
	else
	{
		std::pair<Sci_Position, Sci_Position> v = std::pair<Sci_Position, Sci_Position>(execute(SCI_GETANCHOR), execute(SCI_GETCURRENTPOS));
		pos._selections.emplace_back(v);
	}

	buf->setPosition(pos, this);
}

void ScintillaEditView::restoreCurrentPosPreStep()
{
	Buffer * buf = MainFileManager.getBufferByID(_currentBufferID);
	const Position & pos = buf->getPosition(this);

	execute(SCI_SETSELECTIONMODE, pos._selMode);	//enable
	if (pos._selMode == SC_SEL_STREAM && pos._selections.size() > 1)
	{
		execute(SCI_CANCEL);							//disable
		for (size_t i = 0; i < pos._selections.size(); ++i)
		{
			const Sci_Position startPos = pos._selections[i].first;
			const Sci_Position endPos = pos._selections[i].second;
			if (i == 0)
				execute(SCI_SETSELECTION, endPos, startPos);
			else
				execute(SCI_ADDSELECTION, endPos, startPos);
		}
		execute(SCI_SETMAINSELECTION, pos._selections.size() - 1);
	}
	else
	{
		if (pos._selections.size() > 0)
		{
			const Sci_Position startPos = pos._selections[0].first;
			const Sci_Position endPos = pos._selections[0].second;
			execute(SCI_SETANCHOR, startPos);
			execute(SCI_SETCURRENTPOS, endPos);
		}
		execute(SCI_CANCEL);							//disable
	}
	if (!isWrap()) //only offset if not wrapping, otherwise the offset isnt needed at all
	{
		execute(SCI_SETSCROLLWIDTH, pos._scrollWidth);
		execute(SCI_SETXOFFSET, pos._xOffset);
	}
	execute(SCI_CHOOSECARETX); // choose current x position
	intptr_t lineToShow = execute(SCI_VISIBLEFROMDOCLINE, pos._firstVisibleLine);
	execute(SCI_SETFIRSTVISIBLELINE, lineToShow);
	if (isWrap())
	{
		// Enable flag 'positionRestoreNeeded' so that function restoreCurrentPosPostStep get called
		// once scintilla send SCN_PAITED notification
		_positionRestoreNeeded = true;
	}
	_restorePositionRetryCount = 0;
}

void ScintillaEditView::styleChange()
{
	const bool isSameLangType = _prevBuffer != nullptr && ((_prevBuffer == _currentBuffer) || (_prevBuffer->getLangType() == _currentBuffer->getLangType())); // added
	const int currentLangInt = static_cast<int>(_currentBuffer->getLangType()); // added
	const bool isFirstActiveBuffer = (_currentBuffer->getLastLangType() != currentLangInt) || (_currentBuffer->isUntitled()); // added
	if (!isSameLangType && !isFirstActiveBuffer)  // When entering the tab for the second or more times
	{
		Document prevDoc = execute(SCI_GETDOCPOINTER); // added
		execute(SCI_SETMODEVENTMASK, 0); // added
		execute(SCI_SETDOCPOINTER, 0, getBlankDocument()); // added
		execute(SCI_SETMODEVENTMASK, MODEVENTMASK_ON); // added
		defineDocType(_currentBuffer->getLangType());

		execute(SCI_SETMODEVENTMASK, 0); // added
		execute(SCI_SETDOCPOINTER, 0, prevDoc); // added
		execute(SCI_SETMODEVENTMASK, MODEVENTMASK_ON); // added
	}
	if (isFirstActiveBuffer) // added
	{
		defineDocType(_currentBuffer->getLangType()); // added
	}
	_currentBuffer->setLastLangType(currentLangInt); // added
	restyleBuffer();
}

std::pair<Sci_Position, Sci_Position> ScintillaEditView::getSelectionPosition(intptr_t selectionNumber /* = -1 */, int sort/* = 1*/) const
{
	Sci_Position start_pos, end_pos;

	if ((selectionNumber < 0))
	{
		start_pos = execute(SCI_GETSELECTIONSTART);
		end_pos = execute(SCI_GETSELECTIONEND);
	}
	else
	{
		start_pos = execute(SCI_GETSELECTIONNSTART, selectionNumber);
		end_pos = execute(SCI_GETSELECTIONNEND, selectionNumber);
	}

	if (sort >= 1)
	{
		if (start_pos > end_pos)
			std::swap(start_pos, end_pos);
		return std::pair<Sci_Position, Sci_Position>(start_pos, end_pos);
	}

	if (sort == 0)
	{
		Sci_Position caret_pos = execute(SCI_GETSELECTIONNCARET, selectionNumber);
		if (caret_pos == start_pos)
		{
			std::swap(start_pos, end_pos);
		}
	}

	//If sort <= -1, return directly

	return std::pair<Sci_Position, Sci_Position>(start_pos, end_pos);
}
```


// **Parameters.h**

```
struct Position
{
	intptr_t _firstVisibleLine = 0;
	//deletetd code:  //intptr_t _startPos = 0;
	//deletetd code:  //intptr_t _endPos = 0;
	intptr_t _xOffset = 0;
	intptr_t _selMode = 0;
	intptr_t _scrollWidth = 1;
	intptr_t _offset = 0;
	intptr_t _wrapCount = 0;
	std::vector< std::pair <Sci_Position, Sci_Position> > _selections;  // Save one or more selections
};
```


// **Parameters.cpp**

```
std::vector<std::pair<Sci_Position, Sci_Position>> getSelections(const wchar_t* posStr)
{
	std::vector<std::pair<Sci_Position, Sci_Position>> selections;
	if (posStr == nullptr || *posStr == L'\0')
		return selections;

	const wchar_t* ptr = posStr;

	while (*ptr)
	{
		wchar_t* endPtr = nullptr;
		long long start = std::wcstoll(ptr, &endPtr, 10);
		if (ptr == endPtr)
			break;
		if (*endPtr == L',')
			++endPtr;

		ptr = endPtr;
		long long end = std::wcstoll(ptr, &endPtr, 10);
		if (ptr == endPtr)
			break;
		selections.emplace_back(start, end);

		if (*endPtr == L';')
			++endPtr;
		ptr = endPtr;
	}
	return selections;
}

void setSelections(const std::vector<std::pair<Sci_Position, Sci_Position>>& selections, wchar_t* posStr) // added
{
	if (!posStr)
		return;

	wchar_t* ptr = posStr;

	const size_t selections_size = selections.size();

	if (selections_size == 0)
		return;

	const size_t startSelection = (selections_size <= 99) ? 0 : selections_size - 99;  //20241030:最多只记录99个选区
	for (size_t i = startSelection; i < selections_size; ++i)
	{
		ptr += std::swprintf(ptr, 50, L"%lld,%lld", static_cast<long long>(selections[i].first), static_cast<long long>(selections[i].second));

		if (i != selections_size - 1)
		{
			*ptr++ = L';';
		}
	}
	*ptr = L'\0';
}

bool NppParameters::getSessionFromXmlTree(TiXmlDocument *pSessionDoc, Session& session)
{
	// ...
	for (size_t k = 0; k < nbView; ++k)
	{
		// ...
				if (fileName)
				{
					// ...
					//deletetd code:  //posStr = (childNode->ToElement())->Attribute(L"startPos");
					//deletetd code:  //if (posStr)
					//deletetd code:  //	position._startPos = static_cast<intptr_t>(_ttoi64(posStr));
					//deletetd code:  //posStr = (childNode->ToElement())->Attribute(L"endPos");
					//deletetd code:  //if (posStr)
					//deletetd code:  //	position._endPos = static_cast<intptr_t>(_ttoi64(posStr));
					// ...
					if (posStr)
						position._wrapCount = static_cast<intptr_t>(_ttoi64(posStr));

					posStr = (childNode->ToElement())->Attribute(L"selections"); // added
					if (posStr) // added
						position._selections = getSelections(posStr); // added

					// ...
				}
	}
	// ...
}

void NppParameters::writeSession(const Session & session, const wchar_t *fileName)
{
	// ...
		for (size_t k = 0; k < nbElem ; ++k)
		{
			// ...
			for (size_t i = 0, len = viewElems[k].viewFiles->size(); i < len ; ++i)
			{
				// ...
				//deletetd code:  //(fileNameNode->ToElement())->SetAttribute(L"startPos", _i64tot(static_cast<LONGLONG>(viewSessionFiles[i]._startPos), szInt64, 10));
				//deletetd code:  //(fileNameNode->ToElement())->SetAttribute(L"endPos", _i64tot(static_cast<LONGLONG>(viewSessionFiles[i]._endPos), szInt64, 10));
				// ...
				(fileNameNode->ToElement())->SetAttribute(L"wrapCount", _i64tot(static_cast<LONGLONG>(viewSessionFiles[i]._wrapCount), szInt64, 10));

				wchar_t* selections = new wchar_t[64 * viewSessionFiles[i]._selections.size() + 1] {}; // added
				setSelections(viewSessionFiles[i]._selections, selections); // added
				(fileNameNode->ToElement())->SetAttribute(L"selections", selections); // added
				delete[] selections; // added
				(fileNameNode->ToElement())->SetAttribute(L"lang", viewSessionFiles[i]._langName.c_str());
				// ...
			}
		}
}
```


// **Buffer.h**

```
class Buffer final {
	friend class FileManager;
public:
	// ...
	void setLangType(LangType lang, const wchar_t * userLangName = L"");

	int getLastLangType() const { return _lastLangType; } // added

	void setLastLangType(int val) { _lastLangType = val; } // added

	UniMode getUnicodeMode() const { return _unicodeMode; }
	// ...
private:
	// ...
	LangType _lang = L_TEXT;
	int _lastLangType = -1; // added
	// ...
};
```

// **GoToLineDlg.cpp**

```
intptr_t CALLBACK GoToLineDlg::run_dlgProc(UINT message, WPARAM wParam, LPARAM lParam)
{
	// ...
		case WM_COMMAND:
		{
			switch (LOWORD(wParam))
			{
				// ...
				case IDOK :
                {
					(*_ppEditView)->execute(SCI_SETMODEVENTMASK, 0); // added
                    long long line = getLine();
                    if (line != -1)
                    {
                        display(false);
						if (_mode == go2line)
						{
							// ...
						}
						else
						{
							size_t posToGoto = 0;
							line--; // Subtract 1 from the number on the interface
							if (line > 0)
							// ...
						}
					}
					(*_ppEditView)->execute(SCI_SETMODEVENTMASK, MODEVENTMASK_ON); // added
					// ...
				}
			}
		}
	// ...
}

void GoToLineDlg::updateLinesNumbers() const
{
	// ...
	if (_mode == go2line)
	{
		// ...
	}
	else
	{
		current = (*_ppEditView)->execute(SCI_GETCURRENTPOS) + 1; // Fixed inconsistency with status bar display, need+1 here  //current = (*_ppEditView)->execute(SCI_GETCURRENTPOS);
		size_t currentDocLength = (*_ppEditView)->getCurrentDocLen();
		limit = currentDocLength + 1; // Fixed inconsistency with status bar display, need+1 here  //limit = (currentDocLength > 0 ? currentDocLength - 1 : 0);
	}

	//Resolve the issue of window flickering during the loading process of large files
	char buffer[22] = { 0 }; // added
	::GetDlgItemTextA(_hSelf, ID_CURRLINE, buffer, 21); // added
	if (static_cast<size_t>(std::strtoul(buffer, nullptr, 10)) != current) // added
		::SetDlgItemTextA(_hSelf, ID_CURRLINE, std::to_string(current).c_str());
	::GetDlgItemTextA(_hSelf, ID_LASTLINE, buffer, 21); // added
	if (static_cast<size_t>(std::strtoul(buffer, nullptr, 10)) != limit) // added
		::SetDlgItemTextA(_hSelf, ID_LASTLINE, std::to_string(limit).c_str());
}
```


// **NppIO.cpp**
```
bool Notepad_plus::doReload(BufferID id, bool alert = true)
{
	// ...
	if (mainVisisble)
	{
		_mainEditView.saveCurrentPos();
		_subEditView.execute(SCI_SETMODEVENTMASK, 0); // added
		_mainEditView.execute(SCI_SETDOCPOINTER, 0, 0);
		_subEditView.execute(SCI_SETMODEVENTMASK, MODEVENTMASK_ON); // added
	}

	if (subVisisble)
	{
		_subEditView.saveCurrentPos();
		_subEditView.execute(SCI_SETMODEVENTMASK, 0); // added
		_subEditView.execute(SCI_SETDOCPOINTER, 0, 0);
		_subEditView.execute(SCI_SETMODEVENTMASK, MODEVENTMASK_ON); // added
	}

	// ...
	Buffer * pBuf = MainFileManager.getBufferByID(id);
	pBuf->setLastLangType(-1); // The last language should be initialized here, so that it can be found later in function activateBuffer(id, whichOne);
	if (mainVisisble)
	{
		_mainEditView.execute(SCI_SETMODEVENTMASK, 0); // added
		_mainEditView.execute(SCI_SETDOCPOINTER, 0, pBuf->getDocument());
		_mainEditView.execute(SCI_SETMODEVENTMASK, MODEVENTMASK_ON); // added
		_mainEditView.restoreCurrentPosPreStep();
	}

	if (subVisisble)
	{
		_subEditView.execute(SCI_SETMODEVENTMASK, 0); // added
		_subEditView.execute(SCI_SETDOCPOINTER, 0, pBuf->getDocument());
		_subEditView.execute(SCI_SETMODEVENTMASK, MODEVENTMASK_ON); // added
		_subEditView.restoreCurrentPosPreStep();
	}

	// ...
}
```


// **NppCommands.cpp**

```
void Notepad_plus::command(int id)
{
	switch (id)
	{
        // ...
        case IDM_LANG_USER :
		{
			LangType lang = menuID2LangType(id); // added
			setLanguage(lang);  //setLanguage(menuID2LangType(id));
			// Manually set language, don't change language even file extension changes.
			Buffer *buffer = _pEditView->getCurrentBuffer();
			buffer->langHasBeenSetFromMenu();
			buffer->setLastLangType(static_cast<int>(lang)); // After manually resetting the language, the last language should be set directly to the current language

			if (_pDocMap)
			{
				_pDocMap->setSyntaxHiliting();
			}
		}
        break;
		// ...
	}
}
```


// **NppNotification.cpp**

```
BOOL Notepad_plus::notify(SCNotification *notification)
{
	// ...
	switch (notification->nmhdr.code)
	{
		// ...
		case SCN_MODIFIED:
		{
			// ...
			//deleted code:
			//if (notification->modificationType & SC_MOD_CHANGEINDICATOR)
			//{
			//	::InvalidateRect(notifyView->getHSelf(), NULL, FALSE);
			//}
			break;
		}
		case NM_DBLCLK :
		{
			// ...
				if (lpnm->dwItemSpec == DWORD(STATUSBAR_CUR_POS))
				{
					//deleted code:
					//bool isFirstTime = !_goToLineDlg.isCreated();
					//_goToLineDlg.doDialog(_nativeLangSpeaker.isRTL());
					//if (isFirstTime)
					//	_nativeLangSpeaker.changeDlgLang(_goToLineDlg.getHSelf(), "GoToLine");

					command(IDM_SEARCH_GOTOLINE); // added
				}
			// ...
		}
		case SCN_CHARADDED:
		{
			// ...
				Buffer* currentBuf = _pEditView->getCurrentBuffer();
				if (currentBuf->allowAutoCompletion() && (!currentBuf->isReadOnly()))  //if (currentBuf->allowAutoCompletion())
				{
					// ...
				}
		}
		// ...
	}
}
```


// **Notepad_plus.cpp**
```
…